commands was splitted cogs
