+++ .gitignore
@@ -2,17 +2,6 @@ config.py
 meiryo.ttc
 README.md
 sina_datas.db
-kazuta123-a.png
-kazuta123-b.png
-kazuta123-c.png
-M@ji☆.png
-noimg.png
-sina'slevelcard.png
-tomohiro0405.png
-usericon.png
-weather.png
-雪銀　翔.png
-氷河.png
 musicfile/
 .vscode/
 __pycache__/
