+++ cogs/_template.py
@@ -2,6 +2,7 @@
 
 import discord
 from discord.ext import commands
+import asyncio
 
 import m10s_util as ut
 """↑様々な便利コマンド詰め合わせ
@@ -9,6 +10,12 @@ ut.textto("キー",Member)
     ユーザーの言語設定に基づいてキーのテキストを返す。
 ut.ondevicon(Member)
     オンライン状況に基づくデバイスアイコンテキストを返す。
+ut.getEmbed(title,description,color,(name,value)...)
+    Embedのお手軽生成。これ使ったのがあるから消そうにも消せない。
+await ut.opendm(Member/User)
+    DMチャンネルを返します。DMチャンネルが存在しないなんてことでは困らせません。
+await wait_message_return(ctx,質問するテキスト,←の送信先,待つ時間):
+    入力待ちの簡略化。タイムアウトの例外キャッチを忘れずに
 """
 
 class ClassName(commands.Cog):
