+++ cogs/music.py
@@ -19,9 +19,6 @@ import re
 
 music.py
 制作:mii-10#3110(Discord)
-クレジットをお願いします。
-不具合は上のDiscordアカウントまで。
-予告なく、思惟奈ちゃんで使用しているソースと公開しているソースにずれが生じることがあります。
 """
 
 ytdlopts = {
